refactor(registry): move sync outcome type
