+++ tracecat/registry/actions/types.py
@@ -14,6 +14,14 @@ if TYPE_CHECKING:
     from tracecat.registry.versions.schemas import RegistryVersionManifest
 
 
+@dataclass(frozen=True, slots=True)
+class RepositorySyncOutcome:
+    """Commit and version produced by a repository sync."""
+
+    commit_sha: str | None
+    version: str | None
+
+
 @dataclass(slots=True)
 class IndexEntry:
     """Data holder for registry index entries from UNION ALL query results.
